Move standard components into common package
