+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/webimport/standard/common/domain/DisplayNameFormatter.kt
@@ -1,4 +1,4 @@
-package io.github.composegears.valkyrie.ui.screen.webimport.standard.domain
+package io.github.composegears.valkyrie.ui.screen.webimport.standard.common.domain
 
 import io.github.composegears.valkyrie.parser.unified.ext.capitalized
