+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/webimport/standard/common/model/IconStyle.kt
@@ -1,0 +1,6 @@
+package io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model
+
+data class IconStyle(
+    val id: String,
+    val name: String,
+)
