+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/webimport/standard/common/model/SizeSettings.kt
@@ -1,4 +1,4 @@
-package io.github.composegears.valkyrie.ui.screen.webimport.standard.model
+package io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model
 
 import io.github.composegears.valkyrie.ui.screen.webimport.common.model.IconSettings
