+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/webimport/standard/common/ui/StandardTopActions.kt
@@ -1,4 +1,4 @@
-package io.github.composegears.valkyrie.ui.screen.webimport.standard.ui
+package io.github.composegears.valkyrie.ui.screen.webimport.standard.common.ui
 
 import androidx.compose.foundation.layout.width
 import androidx.compose.runtime.Composable
@@ -10,8 +10,8 @@ import io.github.composegears.valkyrie.jewel.DropdownList
 import io.github.composegears.valkyrie.jewel.tooling.PreviewTheme
 import io.github.composegears.valkyrie.sdk.compose.foundation.rememberMutableState
 import io.github.composegears.valkyrie.ui.screen.webimport.common.ui.WebImportTopActions
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.model.IconStyle
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.model.InferredCategory
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.IconStyle
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.InferredCategory
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
