+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/webimport/standard/lucide/LucideImportScreen.kt
@@ -7,7 +7,7 @@ import com.composegears.tiamat.compose.navDestination
 import com.composegears.tiamat.compose.navigate
 import io.github.composegears.valkyrie.ui.screen.mode.simple.conversion.SimpleConversionParamsSource.TextSource
 import io.github.composegears.valkyrie.ui.screen.mode.simple.conversion.SimpleConversionScreen
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.StandardImportScreen
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.StandardImportScreen
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.lucide.di.LucideModule
 import io.github.composegears.valkyrie.util.stringResource
