--- tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/webimport/standard/model/IconStyle.kt
@@ -1,6 +1,0 @@
-package io.github.composegears.valkyrie.ui.screen.webimport.standard.model
-
-data class IconStyle(
-    val id: String,
-    val name: String,
-)
