+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/webimport/standard/remix/domain/RemixUseCase.kt
@@ -2,15 +2,15 @@ package io.github.composegears.valkyrie.ui.screen.webimport.standard.remix.domai
 
 import io.github.composegears.valkyrie.settings.InMemorySettings
 import io.github.composegears.valkyrie.ui.screen.webimport.common.model.FontByteArray
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.domain.StandardIconProvider
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.domain.SvgSizeCustomizer
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.domain.inferCategoryFromTags
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.domain.toDisplayName
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.model.InferredCategory
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.model.SizeSettings
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.model.StandardIcon
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.model.StandardIconConfig
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.model.toStandardIconConfig
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.domain.StandardIconProvider
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.domain.SvgSizeCustomizer
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.domain.inferCategoryFromTags
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.domain.toDisplayName
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.InferredCategory
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.SizeSettings
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.StandardIcon
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.StandardIconConfig
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.toStandardIconConfig
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.remix.data.RemixRepository
 
 class RemixUseCase(
